ai_scout_batch_2026_04_22/llm-semantic-idempotency-prototype/requirements.txt
openai
pydantic
numpy
scikit-learn
nltk
sentence-transformers
jsonschema
pytest
pytest-mock
python-dotenv
z3-solver